+++ lib/src/models/requests/transcribe_request_dto.g.dart
@@ -23,6 +23,7 @@ _TranscribeRequestDto _$TranscribeRequestDtoFromJson(
       isRealtime: json['is_realtime'] as bool,
       diarize: json['diarize'] as bool,
       speedUp: json['speed_up'] as bool,
+      initialPrompt: json['initial_prompt'] as String?,
     );
 
 Map<String, dynamic> _$TranscribeRequestDtoToJson(
@@ -42,4 +43,5 @@ Map<String, dynamic> _$TranscribeRequestDtoToJson(
       'is_realtime': instance.isRealtime,
       'diarize': instance.diarize,
       'speed_up': instance.speedUp,
+      'initial_prompt': instance.initialPrompt,
     };